+++ contracts/BlockScheduler.sol
@@ -18,5 +18,5 @@ contract BlockScheduler is BaseScheduler {
 }
 
 
-contract TestnetBlockScheduler is BlockScheduler(0x6005cb5aa9c4774c9f1f46ef3323c1337809cdb0) {
-}
+// contract TestnetBlockScheduler is BlockScheduler(0x6005cb5aa9c4774c9f1f46ef3323c1337809cdb0) {
+// }